tests/wpt/tests/content-security-policy/xslt/xslt-strict-dynamic-blocked.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  async_test(t => {
    let iframe = document.createElement("iframe");
    iframe.src = "/content-security-policy/xslt/support/xml-with-stylesheet-strict-dynamic.xml";
    document.body.append(iframe);

    iframe.contentWindow.addEventListener("securitypolicyviolation", t.step_func_done(e => {
        assert_equals(e.effectiveDirective, "script-src-elem");
        assert_true(e.blockedURI.endsWith("/xslt/support/simple.xsl"));
    }));

  }, "XSLT stylesheet is blocked by 'strict-dynamic'");
</script>
</body>
